requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
celery
redis